# docs/swagger.yaml
basePath: /api
definitions:
...
  project-board-api_internal_dto.AddParticipantsRequest:
    description: 'Request to add single or multiple participants to a board For single
      participant: provide array with 1 element For multiple participants: provide
      array with up to 50 elements Duplicate userIds in the request will be automatically
      removed'
    properties:
      boardId:
        example: 1275eac5-f0f9-4bee-8235-576a0042f42b
        type: string
      userIds:
        example:
        - a1b2c3d4-e5f6-7890-abcd-ef1234567890
        - b2c3d4e5-f6a7-8901-bcde-f12345678901
        items:
          type: string
        maxItems: 50
        minItems: 1
        type: array
    required:
    - boardId
    - userIds
    type: object
  project-board-api_internal_dto.AddParticipantsResponse:
    description: 'Response for bulk participant addition HTTP 201: All participants
      added successfully (totalFailed=0) HTTP 207: Partial success (totalSuccess>0
      and totalFailed>0) HTTP 400: All participants failed (totalSuccess=0)'
    properties:
      results:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.ParticipantResult'
        type: array
      totalFailed:
        example: 1
        type: integer
      totalRequested:
        example: 3
        type: integer
      totalSuccess:
        example: 2
        type: integer
    type: object
...
  project-board-api_internal_dto.ParticipantResponse:
    description: Participant information for a board
    properties:
      boardId:
        example: 1275eac5-f0f9-4bee-8235-576a0042f42b
        type: string
      createdAt:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      userId:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
    type: object
  project-board-api_internal_dto.ParticipantResult:
    description: Result for each participant addition attempt success=true means participant
      was added successfully success=false means addition failed, error field contains
      reason
    properties:
      error:
        example: Participant already exists
        type: string
      success:
        example: true
        type: boolean
      userId:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
    type: object
...
  project-board-api_internal_response.ErrorResponse:
    properties:
      error: {}
      requestId:
        type: string
    type: object
  project-board-api_internal_response.SuccessResponse:
    properties:
      data: {}
      requestId:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    프로젝트 보드 관리 시스템 API 서버입니다.
    Board, Project, Comment, Participant 관리 기능을 제공합니다.
    ALB path-based routing with /api/boards prefix
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Project Board Management API
  version: "1.0"
paths:
...
  /participants:
    post:
      consumes:
      - application/json
      description: Board에 한 명 또는 여러 명의 참여자를 추가합니다
      parameters:
      - description: 'Participant 추가 요청 (단건: userIds 배열에 1개, 다중: userIds 배열에 여러 개)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.AddParticipantsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 모든 참여자 추가 성공
          schema:
            $ref: '#/definitions/project-board-api_internal_dto.AddParticipantsResponse'
        "207":
          description: 일부 참여자만 추가 성공 (Multi-Status)
          schema:
            $ref: '#/definitions/project-board-api_internal_dto.AddParticipantsResponse'
        "400":
          description: 잘못된 요청 또는 모든 참여자 추가 실패
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Participant 추가 (단건/다중)
      tags:
      - participants
  /participants/board/{boardId}:
    get:
      description: 특정 Board의 모든 참여자를 조회합니다
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Participant 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.ParticipantResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Board ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board의 Participant 목록 조회
      tags:
      - participants
  /participants/board/{boardId}/user/{userId}:
    delete:
      description: Board에서 참여자를 제거합니다
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      - description: User ID (UUID)
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Participant 제거 성공
          schema:
            $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
        "400":
          description: 잘못된 ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board 또는 Participant를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Participant 제거
      tags:
      - participants
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 입력하세요. 형식: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"